Added score calculation
The more I add the more problems appear with game running. They seem to be memory problems.
I learned that malloc/free-ing without using predefined chunks that we use realloc and controll all the time, will inevitably result in memory fragmentation.
So for now the best solution would be using static constexpr since it will be wrote on the flash, therefore free a lot of RAM memory.